Spring 2024 Assignment 3
* Added shared memory request queue infrastructure to facilitate the implementation of the cryptographic service. A process using this queue will sleep if the queue is empty when attempting to read, or when the queue is full when attempting to write.

* Added the skeleton of the cryptographic server and three system calls: crypto_op() is used by the client to request encryption/decryption from the system; take_shared_memory_request is to be used by a privileged service process such as the cryptographic server to map a request into its address space to work on it; and remove_shared_memory_request removes it from the server's address space.
The cryptographic service uses the request queue to keep track of requests going from clients to the server.
The client requests a decryption of a message, waits for it, then prints the message.
The server should receive a request, map it into its address space, verify that it is valid, then perform encryption/decryption as requested, signal the client that the processing has completed via a flag in the shared memory, then signal the kernel to unmap the request.

* One final piece is protecting the loading of the superblock in the filesystem using a sleeplock, otherwise the kernel is unable to launch multiple processes like init, because they race to load the superblock into memory. This way processes will wait for the first process to load the superblock.

THIS COMMIT IS NOT INTENDED TO COMPILE WITHOUT COMPLETING TASK 1 OF THE ASSIGNMENT!